requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
brotli==1.0.9
    # via flask-compress
click==8.0.3
    # via flask
colorama==0.4.4
    # via click
cycler==0.11.0
    # via matplotlib
dash==2.0.0
    # via -r requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
flask==2.0.2
    # via
    #   dash
    #   flask-compress
flask-compress==1.10.1
    # via dash
fonttools==4.29.0
    # via matplotlib
itsdangerous==2.0.1
    # via flask
jinja2==3.0.3
    # via flask
kiwisolver==1.3.2
    # via matplotlib
markupsafe==2.0.1
    # via jinja2
matplotlib==3.5.1
    # via -r requirements.in
numpy==1.22.1
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
    #   pyarrow
packaging==21.3
    # via matplotlib
pandas==1.4.0
    # via -r requirements.in
pillow==9.0.0
    # via matplotlib
plotly==5.5.0
    # via dash
pyarrow==6.0.1
    # via -r requirements.in
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2021.3
    # via pandas
pyyaml==6.0
    # via -r requirements.in
six==1.16.0
    # via
    #   plotly
    #   python-dateutil
tenacity==8.0.1
    # via plotly
werkzeug==2.0.2
    # via flask
